backend/backend/requirements.txt
asgiref==3.5.2
distlib==0.3.4
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
filelock==3.6.0
gunicorn==20.1.0
numpy==1.22.4
pandas==1.4.2
platformdirs==2.5.1
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
virtualenv==20.14.0
whitenoise==6.2.0
